--- a/Cargo.toml
+++ b/Cargo.toml
@@ -51,6 +51,7 @@ codegen-units = 16
 inherits = "release"
 debug = "full"
 strip = "none"
+panic = "unwind"
 
 [[bench]]
 name = "dcrtpoly"