- split dummy param and real param test for easy build for memory profile 
- hash sampler parallel on index 

## log

real param runs on linux

```
running 1 test
2025-03-24T08:37:18.129083Z  INFO diamond_io::utils: Sampled public data || Current physical/virtural memory usage: 4660285440 | 8555380736
test test::test_io_just_mul_enc_and_bit_real_params has been running for over 60 seconds
2025-03-24T08:44:35.839893Z  INFO diamond_io::utils: Sampled pub key init || Current physical/virtural memory usage: 14326321152 | 18740527104
2025-03-24T08:44:35.992379Z  INFO diamond_io::utils: Sampled s_bars || Current physical/virtural memory usage: 14326452224 | 18740527104
2025-03-24T08:44:36.161305Z  INFO diamond_io::utils: Sampled t_bar_matrix || Current physical/virtural memory usage: 14326452224 | 18740527104
2025-03-24T08:44:36.313762Z  INFO diamond_io::utils: Sampled hardcoded_key_matrix || Current physical/virtural memory usage: 14326452224 | 18740527104
2025-03-24T08:44:37.188701Z  INFO diamond_io::utils: Sampled enc_hardcoded_key_polys || Current physical/virtural memory usage: 14326599680 | 18748981248
2025-03-24T08:44:47.952471Z  INFO diamond_io::utils: Sampled initial encodings || Current physical/virtural memory usage: 15642374144 | 20091158528
2025-03-24T08:45:46.037063Z  INFO diamond_io::utils: b star trapdoor init sampled || Current physical/virtural memory usage: 68480544768 | 75658870784
2025-03-24T08:45:47.950633Z  INFO diamond_io::utils: Computed p_init || Current physical/virtural memory usage: 68521312256 | 75728080896
2025-03-24T08:45:48.539022Z  INFO diamond_io::utils: Computed u_0, u_1, u_star || Current physical/virtural memory usage: 68519231488 | 75548442624
2025-03-24T08:46:41.247456Z  INFO diamond_io::utils: Sampled b_star trapdoor for idx || Current physical/virtural memory usage: 89035173888 | 96663945216
2025-03-24T08:53:59.390943Z  INFO diamond_io::utils: Sampled pub key idx || Current physical/virtural memory usage: 98572627968 | 105388097536
2025-03-24T08:54:00.109976Z  INFO diamond_io::utils: Sampled b_star trapdoor for idx || Current physical/virtural memory usage: 98572627968 | 105388097536
2025-03-24T08:54:47.989888Z  INFO diamond_io::utils: Sampled b trapdoor for idx and bit || Current physical/virtural memory usage: 115310624768 | 121628442624
```

## perf (dummy param)
d=1
```
cargo test --test test_io_dummy_param --no-default-features --features parallel
```

before

```
Time to obfuscate: 15.086361084s
[true, true, true, false]
Time for evaluation: 8.222815833s
Total time: 23.309176917s
```

after

```
Time to obfuscate: 13.375087125s
[true, true, false, false]
Time for evaluation: 7.915870541s
Total time: 21.290957666s
```
---

d=3

before
```
Time to obfuscate: 85.061848833s
[true, true, false, false]
Time for evaluation: 34.450563875s
Total time: 119.512412708s
```

after
```
Time to obfuscate: 84.350289208s
[false, true, true, true]
Time for evaluation: 31.66687625s
Total time: 116.017165458s
```