@enricobottazzi 2025-03-24
Can you set `d=3` and measure performance and memory consumption improvement (if any) ?